Change names of services to CartObject, TicketObject and CheckoutService
