+++ tutorials/tour-of-restate-java/src/main/java/dev/restate/tour/part1/TicketObject.java
@@ -9,23 +9,25 @@
  * https://github.com/restatedev/examples/
  */
 
-package dev.restate.tour.part2;
+package dev.restate.tour.part1;
 
 import dev.restate.sdk.ObjectContext;
 import dev.restate.sdk.annotation.Handler;
 import dev.restate.sdk.annotation.VirtualObject;
 
-import java.time.Duration;
-
 @VirtualObject
-public class TicketService {
+public class TicketObject {
 
     // <start_reserve>
     @Handler
     public boolean reserve(ObjectContext ctx) {
-        //good-code-start
-        ctx.sleep(Duration.ofSeconds(65));
-        //good-code-end
+        //bad-code-start
+        try {
+            Thread.sleep(65000);
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        }
+        //bad-code-end
 
         return true;
     }
