+++ tutorials/tour-of-restate-java/src/main/java/dev/restate/tour/part5/CartObject.java
@@ -9,7 +9,7 @@
  * https://github.com/restatedev/examples/
  */
 
-package dev.restate.tour.part4;
+package dev.restate.tour.part5;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import dev.restate.sdk.ObjectContext;
@@ -20,14 +20,13 @@ import dev.restate.sdk.serde.jackson.JacksonSerdes;
 import dev.restate.tour.app.CheckoutClient;
 import dev.restate.tour.app.UserSessionClient;
 import dev.restate.tour.auxiliary.CheckoutRequest;
-import dev.restate.tour.part5.TicketServiceClient;
 
 import java.time.Duration;
 import java.util.HashSet;
 import java.util.Set;
 
 @VirtualObject
-public class UserSession {
+public class CartObject {
 
     public static final StateKey<Set<String>> STATE_KEY = StateKey.of("tickets", JacksonSerdes.of(new TypeReference<>() {}));
 
@@ -37,13 +36,13 @@ public class UserSession {
         boolean reservationSuccess = TicketServiceClient.fromContext(ctx, ticketId).reserve().await();
 
         if (reservationSuccess) {
-            var tickets = ctx.get(STATE_KEY).orElseGet(HashSet::new);
+            Set<String> tickets = ctx.get(STATE_KEY).orElseGet(HashSet::new);
             tickets.add(ticketId);
             ctx.set(STATE_KEY, tickets);
 
             UserSessionClient.fromContext(ctx, ctx.key())
-                    .send(Duration.ofMinutes(15))
-                    .expireTicket(ticketId);
+                .send(Duration.ofMinutes(15))
+                .expireTicket(ticketId);
         }
 
         return reservationSuccess;
@@ -61,7 +60,6 @@ public class UserSession {
         }
     }
 
-    // <start_checkout>
     @Handler
     public boolean checkout(ObjectContext ctx) {
         Set<String> tickets = ctx.get(STATE_KEY).orElseGet(HashSet::new);
@@ -75,16 +73,12 @@ public class UserSession {
                 .await();
 
         if (checkoutSuccess) {
-            //highlight-start
             tickets.forEach(t ->
-                    TicketServiceClient.fromContext(ctx, t).send().markAsSold()
+                TicketServiceClient.fromContext(ctx, t).send().markAsSold()
             );
-            //highlight-end
             ctx.clear(STATE_KEY);
         }
 
         return checkoutSuccess;
     }
-    // <end_checkout>
-
 }
